requirements.txt
anthropic>=0.39.0
supabase>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
